Move TransformCommandLabels from lib/web/ui to lib/ui
