--- lib/web/ui/labels.go
@@ -1,33 +1,0 @@
-/*
- * Teleport
- * Copyright (C) 2023  Gravitational, Inc.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package ui
-
-import (
-	"github.com/gravitational/teleport/api/types"
-)
-
-func transformCommandLabels(commandLabels map[string]types.CommandLabel) map[string]string {
-	labels := make(map[string]string, len(commandLabels))
-
-	for name, cmd := range commandLabels {
-		labels[name] = cmd.GetResult()
-	}
-
-	return labels
-}
