--- lib/web/ui/labels_test.go
@@ -1,41 +1,0 @@
-/*
- * Teleport
- * Copyright (C) 2023  Gravitational, Inc.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package ui
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-
-	"github.com/gravitational/teleport/api/types"
-)
-
-func TestTransformCommandLabels(t *testing.T) {
-	commandLabels := map[string]types.CommandLabel{
-		"label1": &types.CommandLabelV2{
-			Result: "value1",
-		},
-	}
-	labels := transformCommandLabels(commandLabels)
-	expected := map[string]string{
-		"label1": "value1",
-	}
-
-	require.Equal(t, expected, labels)
-}
